requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
cffi==1.11.5
coverage==4.5.2
cryptography==2.4.2
decorator==4.3.0
Django==2.1.3
django-extensions==2.1.4
djangorestframework==3.9.0
idna==2.7
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.0
pytz==2018.7
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
